backend/textileflow/urls.py
"""
URL configuration for textileflow project.
"""
from django.contrib import admin
from django.urls import path, include
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    path('admin/', admin.site.urls),
    
    # Authentication
    path('api/auth/login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/auth/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    
    # App URLs
    path('api/', include('core.urls')),
    path('api/', include('inventory.urls')),
    path('api/', include('transactions.urls')),
    path('api/', include('accounts.urls')),
]
